Move Version subclasses into their own files
