+++ Core/CKAN-core.csproj
@@ -110,6 +110,7 @@
     <Compile Include="ServiceLocator.cs" />
     <Compile Include="Types\CkanModule.cs" />
     <Compile Include="Types\CKANVersion.cs" />
+    <Compile Include="Types\DllVersion.cs" />
     <Compile Include="Types\ExportFileType.cs" />
     <Compile Include="Types\FileType.cs" />
     <Compile Include="Types\GameComparator\BaseGameComparator.cs" />
@@ -121,6 +122,7 @@
     <Compile Include="Types\License.cs" />
     <Compile Include="Types\ModuleInstallDescriptor.cs" />
     <Compile Include="Types\ModuleResolution.cs" />
+    <Compile Include="Types\ProvidesVersion.cs" />
     <Compile Include="Types\ReleaseStatus.cs" />
     <Compile Include="Types\Repository.cs" />
     <Compile Include="Types\Version.cs" />
