+++ Core/Types/DllVersion.cs
@@ -1,0 +1,17 @@
+namespace CKAN
+{
+    /// <summary>
+    /// This class represents a DllVersion. They don't have real
+    /// version numbers or anything
+    /// </summary>
+    public class DllVersion : Version {
+        public DllVersion() :base("0")
+        {
+        }
+
+        override public string ToString()
+        {
+            return AutodetectedDllString;
+        }
+    }
+}
