+++ Core/Types/ProvidesVersion.cs
@@ -1,0 +1,20 @@
+namespace CKAN
+{
+    /// <summary>
+    /// This class represents a virtual version that was provided by
+    /// another module.
+    /// </summary>
+    public class ProvidesVersion : Version {
+        internal readonly string provided_by;
+
+        public ProvidesVersion(string provided_by) :base("0")
+        {
+            this.provided_by = provided_by;
+        }
+
+        override public string ToString()
+        {
+            return string.Format("provided by {0}", provided_by);
+        }
+    }
+}
