+++ Core/Types/Version.cs
@@ -331,40 +331,5 @@ namespace CKAN
         {
             return v1.CompareTo(v2) >= 0;
         }
-
-    }
-
-    /// <summary>
-    /// This class represents a DllVersion. They don't have real
-    /// version numbers or anything
-    /// </summary>
-    public class DllVersion : Version {
-        public DllVersion() :base("0")
-        {
-        }
-
-        override public string ToString()
-        {
-            return AutodetectedDllString;
-        }
-    }
-
-    /// <summary>
-    /// This class represents a virtual version that was provided by
-    /// another module.
-    /// </summary>
-    public class ProvidesVersion : Version {
-        internal readonly string provided_by;
-
-        public ProvidesVersion(string provided_by) :base("0")
-        {
-            this.provided_by = provided_by;
-        }
-
-        override public string ToString()
-        {
-            return string.Format("provided by {0}", provided_by);
-        }
     }
 }
-
